Some more requested changes and hopefully a compiling WinUI without me needing to boot windows

+++ Sources/SwiftCrossUI/Values/ButtonStyle.swift
@@ -1,7 +1,7 @@
 /// A type that applies standard interaction behavior and a custom appearance to all buttons within a view hierarchy.
 public struct ButtonStyle: Hashable, Sendable {
     // ALWAYS KEEP ORDER CONSISTENT with CustomButton.ButtonStyle in Kotlin.
-    package enum Kind: Int {
+    package enum Kind: Int16 {
         case bordered
         case plain
         case borderless
